Keep the label in position once it is shown
When the scroller has a few rows it is possible to perform a jump scroll followed by some cont scrolls. In this situation the label keeps anchored inside the scrolled div to the last position where the jump happened.

Keeping the visibility status we can adjust the position of the label also in the following cont scrolls.

Sync to source repo @1e99ff82642d29fd9a77cee64d78a5cfa299b922